// src/types.ts
export type SolanaNetwork = 'mainnet-beta' | 'devnet' | 'testnet';
export type LitSolanaChain = 'solana' | 'solanaDevnet' | 'solanaTestnet';